backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
aiohttp==3.9.1
pydantic[email]==2.5.0
python-jose[cryptography]==3.3.0
python-multipart==0.0.6

